Refactor checksum calculation
